+++ src/inspect_ai/_cli/eval.py
@@ -159,10 +159,16 @@ def eval_options(func: Callable[..., Any]) -> Callable[..., click.Context]:
         envvar="INSPECT_EVAL_MAX_SUBPROCESSES",
     )
     @click.option(
-        "--max-messages",
+        "--message-limit",
         type=int,
-        help="Maximum number of messages to allow in a task conversation.",
-        envvar="INSPECT_EVAL_MAX_MESSAGES",
+        help="Limit on total messages used for each sample.",
+        envvar="INSPECT_EVAL_MESSAGE_LIMIT",
+    )
+    @click.option(
+        "--token-limit",
+        type=int,
+        help="Limit on total tokens used for each sample.",
+        envvar="INSPECT_EVAL_TOKEN_LIMIT",
     )
     @click.option(
         "--fail-on-error",
@@ -359,7 +365,8 @@ def eval_command(
     parallel_tool_calls: bool | None,
     max_tool_output: int | None,
     cache_prompt: str | None,
-    max_messages: int | None,
+    message_limit: int | None,
+    token_limit: int | None,
     max_samples: int | None,
     max_tasks: int | None,
     max_subprocesses: int | None,
@@ -376,13 +383,14 @@ def eval_command(
     config = config_from_locals(dict(locals()))
 
     # resolve common options
-    (log_dir, log_level) = resolve_common_options(kwargs)
+    (log_dir, log_level, log_level_transcript) = resolve_common_options(kwargs)
 
     # exec eval
     eval_exec(
         tasks=tasks,
         solver=solver,
         log_level=log_level,
+        log_level_transcript=log_level_transcript,
         log_dir=log_dir,
         model=model,
         model_base_url=model_base_url,
@@ -396,7 +404,8 @@ def eval_command(
         epochs=epochs,
         epochs_reducer=epochs_reducer,
         limit=limit,
-        max_messages=max_messages,
+        message_limit=message_limit,
+        token_limit=token_limit,
         max_samples=max_samples,
         max_tasks=max_tasks,
         max_subprocesses=max_subprocesses,
@@ -496,7 +505,8 @@ def eval_set_command(
     parallel_tool_calls: bool | None,
     max_tool_output: int | None,
     cache_prompt: str | None,
-    max_messages: int | None,
+    message_limit: int | None,
+    token_limit: int | None,
     max_samples: int | None,
     max_tasks: int | None,
     max_subprocesses: int | None,
@@ -515,13 +525,14 @@ def eval_set_command(
     config = config_from_locals(dict(locals()))
 
     # resolve common options
-    (log_dir, log_level) = resolve_common_options(kwargs)
+    (log_dir, log_level, log_level_transcript) = resolve_common_options(kwargs)
 
     # exec eval
     success = eval_exec(
         tasks=tasks,
         solver=solver,
         log_level=log_level,
+        log_level_transcript=log_level_transcript,
         log_dir=log_dir,
         model=model,
         model_base_url=model_base_url,
@@ -535,7 +546,8 @@ def eval_set_command(
         epochs=epochs,
         epochs_reducer=epochs_reducer,
         limit=limit,
-        max_messages=max_messages,
+        message_limit=message_limit,
+        token_limit=token_limit,
         max_samples=max_samples,
         max_tasks=max_tasks,
         max_subprocesses=max_subprocesses,
@@ -564,6 +576,7 @@ def eval_exec(
     tasks: tuple[str] | None,
     solver: str | None,
     log_level: str,
+    log_level_transcript: str,
     log_dir: str,
     model: str,
     model_base_url: str | None,
@@ -577,7 +590,8 @@ def eval_exec(
     epochs: int | None,
     epochs_reducer: str | None,
     limit: str | None,
-    max_messages: int | None,
+    message_limit: int | None,
+    token_limit: int | None,
     max_samples: int | None,
     max_tasks: int | None,
     max_subprocesses: int | None,
@@ -638,12 +652,14 @@ def eval_exec(
             sandbox=parse_sandbox(sandbox),
             sandbox_cleanup=sandbox_cleanup,
             log_level=log_level,
+            log_level_transcript=log_level_transcript,
             log_dir=log_dir,
             limit=eval_limit,
             epochs=eval_epochs,
             fail_on_error=fail_on_error,
             debug_errors=debug_errors,
-            max_messages=max_messages,
+            message_limit=message_limit,
+            token_limit=token_limit,
             max_samples=max_samples,
             max_tasks=max_tasks,
             max_subprocesses=max_subprocesses,
@@ -792,7 +808,7 @@ def eval_retry_command(
 ) -> None:
     """Retry failed evaluation(s)"""
     # resolve common options
-    (log_dir, log_level) = resolve_common_options(kwargs)
+    (log_dir, log_level, log_level_transcript) = resolve_common_options(kwargs)
 
     # resolve negating options
     sandbox_cleanup = False if no_sandbox_cleanup else None
@@ -809,6 +825,7 @@ def eval_retry_command(
     eval_retry(
         retry_log_files,
         log_level=log_level,
+        log_level_transcript=log_level_transcript,
         log_dir=log_dir,
         max_samples=max_samples,
         max_tasks=max_tasks,